mac80211: don't remove tainted keys after not programming
When a key is tainted during resume, it is no longer programmed
into the device; however, it's uploaded flag may (will) be set.
Clear the flag when not programming it because it's tainted to
avoid attempting to remove it again later.

Change-Id: Iab301addd7022da539e8bed9b8cac978ea3eccb2
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/45490
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: a94b5f711cac001b4d505626f134a3e2ddf9a392