iwlwifi: sdio: zero trans->hw_rev before reading it
In 8000 HW family starting B-step there is an optimization
for reading in SDIO without having to later clear the 0x8
register (IWL_SDIO_INTR_CAUSE_REG). This already exists in
the code and works for the INIT image, but when switching to
OPERATIONAL the opt. causes an issue, since the opt. is only
partially set at this stage.

Zeroing out trans->hw_rev before trying to read it disables
this opt. just for this one read (by clearing the 0x8 register
regardless of the step) and fixes the issue, allowing later
reads to be done when the op. is completely set.

Change-Id: Ief988ee3ef87465f5a9665336f21c0d148967eee
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/46514
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 8de164eae936fe274e963fd433877e79b9db28cc