iwlwifi: sdio: use correct fw file in 8000 b-step
In 8000 B-step the FW file has changed, but by the time we
know the HW step in SDIO the FW file is already requested.

This patch defaults 8000 family to B-step if no HW step is
detected in time. When it can it checks what HW step it
really is (in 8000 family) and if it isn't B-step, the FW
file is released and the A-step file is requested.

Change-Id: I7b89e9e0827d5425ba3c5c642ebbf2ce80cb4bb9
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/46448
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 5b39d0e5be784dcce8e37543d62e56acf9c59aa1