iwlwifi: mvm: don't capture firmware coredump for D3->D0 reconfig [BUGFIX]
The code to capture firmware errors works during the reconfiguration
phase after an error. As the D3->D0 transition uses the same flow to
get the D0 image reconfigured, this triggered and caused a firmware
coredump to be collected. This in turn, if it isn't picked up by
userspace, can cause module unloading to fail, which is how the bug
was detected.

To fix this issue, introduce a new status flag (D3_RECONFIG) and use
it to detect that during reconfiguration no coredump should be taken
and reported.

Change-Id: I501965d6639122e0587bb300fec29b8709a0a57f
Reported-by: Avi Kraif <avix.kraif@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
MWG100221130
Reviewed-on: https://gerrit.rds.intel.com/r/46751
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: bf692ff491f8b836bc2ced024236e6938b8ffc22