mac80211: ask driver to look at power level when starting AP
The power level might have been set, but as the interface was idle
it might not have taken effect yet. Ask the driver to check the
power level when starting up an AP so that in this case the correct
power level is used in case the device/driver can only set it when
the interface is actually active.

Change-Id: I2e0ed171ce3d60bc4c06cf30b615e25a9038b643
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
MWG100221668
Reviewed-on: https://gerrit.rds.intel.com/r/47148
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: 3d94df97b201fc3eca56b4a5f95380e6781bb4ce