[SQUASH] iwlwifi: mvm: fix missing padding
In a very unintuitive way, mac_addr_u in the firmware code is a union
of 8 bytes, not ETH_ALEN as one might expect.  This caused me to miss
a 2 bytes padding in the iwl_scan_offload_profile_match structure.
Fix it.

Additionally, use the opportunity to reduce the name of the
matching_channels length macro.

This should be squashed with commit 22feefa766a5 (iwlwifi: mvm: check
and report if wake up was due to net detect).

Change-Id: Idb88f35000eae6e7366f1058a705b072166e9666
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/48038
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 0fbda6d914ab9a00b1773bd9b803e07070dd6a60