iwlwifi/backports integration: add BACKPORTS_BUILD/BRANCH_TSTAMP
The new BACKPORTS_BUILD_TSTAMP will default to '__DATE__ " " __TIME__'
which allows us to determine when the build was made. Put this into
the 'versions' file instead of the C source directly to allow for it
to be overridden when creating Core/Ubuntu builds.

The new BACKPORTS_BRANCH_TSTAMP is unset for builds off master and
will be set by the prune script (change also made here) to the merge
base between the branch and master.

Change-Id: Ib8ba213f5f5a6ae59751eb78cddaefdf97234425
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/48073
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: 6906c936adfaa3e055e129b1f64b0615e3727378