iwlwifi: pcie: re-ACK all interrupts after device reset [BUGFIX]
When we reset the device, the CSR_INT gets cleared as well
as CSR_INT_MASK. Meaning that we shouldn't get any interrupt
but, due to a hardware bug, recent devices will keep sending
interrupts. This leads to an interrupt storm while stopping
the device.
The way to fix this is to ACK all the interrupts after the
device is reset so that the value of CSR_INT will stay
0xffffffff.

Change-Id: I7684e2959aa54af32d2519ff2ee2290bb6461dc6
Fixes: 522713c81e4e ("iwlwifi: pcie: properly reset the device")
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
MWG100223549
Reviewed-on: https://gerrit.rds.intel.com/r/48555
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: f7b88886e54d4fd3e7e11ff2667a2f90ee24f8c2