iwlwifi: mvm: BT Coex - make the primary / secondary pick traffic aware [NOUPSTREAM]
The primary channel is the channel that will be untouched
by BT. The secondary channel might be touched by BT. Hence,
we want the primary to be the most active channel.
To do so, use the TCM infrastructure.
Since the BT keeps sending notifications, we can rely on
them to trigger the check. Every 10 seconds, we will check
what is the most active context and chose the right
primary.
We need to wait 10 seconds before we modify the settings
because frequent changes in these settings can confuse BT.

Change-Id: I0cc0ff9762fd416c69b5079f75532441c8afbde4
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/49070
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 1afc46757019a964196f6596b2bb6938387c0c16