[NOUPSTREAM]: mac80211: align ieee80211_ch_switch_post_beacon() with upstream
This is a bogus change just to align with upstream.  Upstream is
wrong, the cfg80211_ch_switch_notify() call should be at the end of
the function, so this patch is wrong, but will make it easier to
handle if a patch to fix upstream is applied on top of this.

Change-Id: I909f557b3aff4a06ab4a865b62d43002fa06fbce
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/49407
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: ab05665e1bf2720d0f1a3e23e1b81f96ed6eafa8