[BUGFIX] iwlwifi: mvm: make mvmvif->mvm always valid
This struct member is currently only valid under debugfs,
but for work structs in mvmvif it is always needed. Make
it always available and fill it earlier as well.

Change-Id: Ie01028bda25de1f91f0f6bd1b08aa667d520eb9a
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
In particular, this fixes a U-APSD compilation issue when
DEBUGFS is not configured, which Benzi reported.
Reviewed-on: https://gerrit.rds.intel.com/r/49999
Reviewed-by: Dor Shaish <dor.shaish@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: bcc21b7bbf7ef95b771b4a74d7c7736c449a46df