iwlwifi: mvm: tcm: account rx ampdu airtime only once
The fw passes the airtime in the phyinfo. In case of
aggregation, the airtime is calculated for entire
amdpu, but the driver increases the airtime (by this
value) for each frame. This results in overly excessive rx
airtime, leading to wrong quota calculations, etc.

Remember the last encountered ampdu_ref (which is
increasing counter), and increase the airtime only
if this is a new one.

Change-Id: Iaafea50121608ee5c73dcca85657ffc5f3e24dd3
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/50156
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 8f0677013a9870f36dcca0158c35f67a3feb3942