iwlwifi: mvm: Add New Tx Power api (SAR DCR)
This new api supports a slightly different behavior that is now expected
from the driver:
 - Default MAC and Device is no restriction.
 - AUX MAC is not expected to be set (factored under device limit).
   The corresponding FW logic: if MAC limit exists, FW will take
   MIN(MAC limit, Device limit), if MAC limit doesn't exist,
   (not set by driver or non-MAC tx like scan) then FW will take the device
   limit.
   Note: Device limit is anyway factored with internal FW restrictions
   such as regulatory, EVM and BT\LTE coex inputs.
When mac80211 configure the hw with new Tx power level, use the new
API to inform the fw with that Tx power limitation per device.
Do not compromise the old scheme in which this is sent per vif as well,
our fw takes the min value anyway.

Change-Id: I360e7e6d2b77d1eab01ee8f3a7bd3bf720d77fa3
Signed-off-by: Avri Altman <avri.altman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/48976
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 56f132813289e5aacb5304e4589a5ddf00085e98