iwlwifi: mvm: rs: use STBC regardless of power save mode
Tx STBC was used only when in CAM mode or if powersave is disabled.
Effectively this meant we never used STBC as these modes aren't
used on most platforms by default. Change that.

Change-Id: Ib16209a3d4310c85a2d44c74285bfacc3c516fb5
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
We want STBC to be enabled by default on Android/Chrome where
we are benchmarked by default not in CAM.
This is per Marina's request.
Reviewed-on: https://gerrit.rds.intel.com/r/50759
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 4f690881ecad3c82f21ccbca42609f01af06f831