iwlwifi: mvm: send the SCD_CFG in SYNC mode when possible [SQUASH]
Sending the SCD_CFG in ASYNC led to a race in the down flow.
We were sending 4 commands to the firmware to de-configure
all the queues and without waiting for them to be handled by
the firwmare we were continuing the down flow.
This led to a situation where we unmapped the host command
queue, but then, the command response came and their was
not corresponding command sent leading to the following
print:

iwl_pcie_cmdq_reclaim: Read index for DMA queue txq id (9), index 40 is out of range [0-256] 41 41

fix that.

Change-Id: Ibc6699dd3ff4c634fff0335fc2f7570275912c6b
Fixes: b0636e505987 ("iwlwifi: mvm: let the firmware configure the scheduler")
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/50886
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: 03c484e02f08118a9179377f854f16f2fb9543e7