iwlwifi: pcie: let the driver set the SCD WrPtr [SQUASH]
The shadow is a mirror of the Write Pointer (WrPtr) of the
scheduler. The shadow register has the advantage to be in
the always-on zone which means that the driver doesn't need
to wake up the device before updating it.

The new way to initialize a Tx queue is to let the firmware
do all the work. But if the firmware updates the WrPtr in
the registers of the scheduler itself, we may have a
situation in which the shadow register contains garbage and
the "real data" is in the registers of the scheduler. This
breaks the assumption based on which the hardware was design
and in that case the garbage from the shadow will be copied
into the scheduler's WrPtr. Since the firmware can't access
the shadow register, fix this by setting the shadow register
from the driver before sending the host command.

Change-Id: I6a642b8df85494b71c379d4da65c7759cdb74f3c
Fixes: b0636e505987 ("iwlwifi: mvm: let the firmware configure the scheduler")
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/50942
x-iwlwifi-stack-dev: 415f54a231b80bb2f6ade11b01f35030f370f2b5