iwlwifi: mvm: ignore temperature updates in the RX statistics notification [BUGFIX]
If the firmware sends spontaneous DTS notfications with the
temperature (indicated in a TLV), we can ignore the temperature we get
in the RX statistics notifications.  This prevents potentially
handling the same temperature change twice.  It also ignores
notifications with temperature equal to 0 that happens from time to
time.

Change-Id: I9d72f4298cb486af116a0193bd9ade97d14f095d
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---
MWG100216743 (partially, need a firmware with the proper TLV as well)
Reviewed-on: https://gerrit.rds.intel.com/r/50962
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: bbab8a53ca4e59397712891eb52d36e21df1353e