mac80211: only roll back station states for WDS when suspending
In the suspend code, we shouldn't roll back the station states because
this is taken care of in other parts of the code, except for WDS
interfaces.  For AP types of interfaces, cfg80211 takes care of
disconnecting all stations before calling the driver's suspend code.
For station interfaces, this will be done in the quiesce code.

For WDS interfaces we still need to do it here, so move the code into
a new case switch for WDS.

Change-Id: I2145002ee46319aa574e10a12aa5dd4f949b1427
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---
MWG MWG100227218
Reviewed-on: https://gerrit.rds.intel.com/r/51170
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: e527c172c289a961c9f25673aba52ca665a171dc